feat: Add DropdownSelector component and shadow modifier
This commit introduces a new `DropdownSelector` composable component. This component displays a label and a button styled to look like a dropdown input field. It allows for custom leading and trailing content.

Additionally, a `dropShadow` modifier extension has been added to apply a customizable drop shadow effect to composables. This includes options for shape, color, blur, offset, and spread.

The `AppColors` have been updated to include `Shadow` and `Shadow_SL08` for use with the shadow modifier.

Finally, the `applyIf` modifier extension and a `verticalColumnScrollbar` modifier for adding a customizable vertical scrollbar have been included.